requirements.txt
fastapi==0.109.2
python-multipart==0.0.9
openai-whisper==20231117
uvicorn==0.27.1 